third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/access-control-preflight-sync-header-denied-expected.txt

CONSOLE WARNING: line 17: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE ERROR: line 35: Failed to load http://localhost:8000/xmlhttprequest/resources/access-control-preflight-denied-xsrf.php?state=header: Request header field X-NON-STANDARD is not allowed by Access-Control-Allow-Headers in preflight response.
PASS: Request successfully blocked.
